Fix race condition when unhooking client
What the original TODO comment failed to realize was that
in debug builds, LLVM codegens an access to ORIGINAL_CLIENT_FUNCS.
So the following scenario caused the game to crash:
1. client::Hook's drop() called.
2. One of the hooked client functions is about to yank from ORIGINAL_CLIENT_FUNCS.
3. drop() moves the client functions out of the Option, leaving a None behind.
4. The hooked client function yanks from None.

Since yank() assumes the Option is always the Some variant,
we encounter immediate UB when trying to yank() on a None variant.

In this case, the "undefined behavior" is the game crashing.

"But wait, the hooked client functions are single-threaded.
How can one of the hooked client functions race
client::Hook::drop()?"

The hooked client functions occupy a single thread.
We call the drop() function from the thread we created
in DllMain().

So there are two threads to keep in mind at all times:
1. The game thread (accesses hooked functions).
2. Our thread (manages hooking and unhooking).