Warn on pedantic clippy lints
I disabled the `find_map` lint because that's a stylistic lint that I
don't agree with.

I disabled the `unseparated_literal_suffix` lint on the generated
bindings to the HL1 SDK because there's a ton of them. Bindgen chose
one style, and we don't have control over that style, so we disable
this stylistic lint as well.

I disabled the `cast_possible_truncation` lint on one item in
module.rs because clippy complained that casting a usize to a
u32 is lossy on platforms where usize is 64 bits. Since we are only
compiling our binary as 32 bits, the conversion between usize and u32
is lossless, and this lint doesn't apply to the relevant code.

I disabled the `cast_possible_truncation` and `cast_possible_wrap`
lints on the `print_text` virtual function of Surface. This lint
correctly point out that if we pass in a wide-text array with more than
i32::MAX elements, the length argument will wrap into negative values.
Instead of disabling this lint, I could have added a bounds check
prior to calling the virtual function, but I chose not to because it is
very unlikely we'll be creating a ~4.3 GB string to trigger
the integer wrap here.